rabbitmq/ssl_requirements.txt
certifi==2020.12.5
pika==1.1.0
python-dotenv==0.15.0
